[Silver III] Title: 주유소, Time: 440 ms, Memory: 36180 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -7,24 +7,24 @@ public class Main {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		int n = Integer.parseInt(br.readLine());
 		StringTokenizer st = new StringTokenizer(br.readLine());
-		int[] distance = new int[n - 1];
-		int[] price = new int[n];
+		long[] distance = new long[n - 1];
+		long[] price = new long[n];
 		
 		for (int i = 0; i < distance.length; i++) {
-			distance[i] = Integer.parseInt(st.nextToken());
+			distance[i] = Long.parseLong(st.nextToken());
 		}
 		
 		st = new StringTokenizer(br.readLine());
 		for (int i = 0; i < price.length; i++) {
-			price[i] = Integer.parseInt(st.nextToken());
+			price[i] = Long.parseLong(st.nextToken());
 		}
 		
-		int lowPrice = price[0];
+		long lowPrice = price[0];
 		long sum = distance[0] * lowPrice;
 		
 		for (int i = 1; i < n - 1; i++) {
 			lowPrice = Math.min(lowPrice, price[i]);
-			sum += (long) distance[i] * lowPrice;
+			sum += distance[i] * lowPrice;
 		}
 		
 		System.out.print(sum);
